Move pauli word logic from `qml.grouping.utils.is_commuting` to `qml.is_commuting` (#3033)
* move pauli word is_commuting to qml.is_commuting

* clean up test file

* no try-except with improved is_pauli_word, update docstring

* update changelog

* move changelog entry to breaking changes

* remove stale "Raises TypeError" from docstring
